+++ app/src/main/java/com/sunasterisk/anticovid_19/base/BaseView.kt
@@ -1,0 +1,5 @@
+package com.sunasterisk.anticovid_19.base
+
+interface BaseView {
+
+}
