+++ app/src/main/java/com/sunasterisk/anticovid_19/data/resource/CovidDataSource.kt
@@ -1,0 +1,14 @@
+package com.sunasterisk.anticovid_19.data.resource
+
+import com.sunasterisk.anticovid_19.data.model.Country
+import com.sunasterisk.anticovid_19.data.resource.remote.util.OnDataLoadCallBack
+
+interface CovidDataSource {
+    interface Local {
+
+    }
+
+    interface Remote {
+        fun getCountryInformation(callback: OnDataLoadCallBack<List<Country>>)
+    }
+}
