+++ app/src/main/java/com/sunasterisk/anticovid_19/data/resource/remote/util/OnDataLoadCallBack.kt
@@ -1,0 +1,8 @@
+package com.sunasterisk.anticovid_19.data.resource.remote.util
+
+import java.lang.Exception
+
+interface OnDataLoadCallBack<T> {
+    fun onSuccess(data: T)
+    fun onFail(exception: Exception)
+}
