+++ app/src/main/java/com/sunasterisk/anticovid_19/utils/Constants.kt
@@ -1,0 +1,18 @@
+package com.sunasterisk.anticovid_19.utils
+
+object ModelConst {
+    const val NEW_CONFIRMED = "NewConfirmed"
+    const val NEW_DEATHS = "NewDeaths"
+    const val NEW_RECOVERED = "NewRecovered"
+    const val TOTAL_CONFIRMED = "TotalConfirmed"
+    const val TOTAL_DEATHS = "TotalDeaths"
+    const val TOTAL_RECOVERED = "TotalRecovered"
+}
+
+object NameConst {
+    const val COUNTRIES = "Countries"
+}
+
+object LinkConst {
+    const val COVID_API = "https://api.covid19api.com/summary"
+}
