--- app/src/main/java/sun/trainingcourse/anticovid_19/base/BaseActivity.kt
@@ -1,17 +1,0 @@
-package sun.trainingcourse.anticovid_19.base
-
-import android.os.Bundle
-import androidx.annotation.LayoutRes
-import androidx.appcompat.app.AppCompatActivity
-
-abstract class BaseActivity : AppCompatActivity() {
-
-    @get:LayoutRes
-    protected abstract val layoutResource: Int
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(layoutResource)
-    }
-
-}
