--- app/src/main/java/sun/trainingcourse/anticovid_19/base/BaseView.kt
@@ -1,5 +1,0 @@
-package sun.trainingcourse.anticovid_19.base
-
-interface BaseView {
-
-}
