--- app/src/main/java/sun/trainingcourse/anticovid_19/data/resource/CovidDataSource.kt
@@ -1,11 +1,0 @@
-package sun.trainingcourse.anticovid_19.data.resource
-
-interface CovidDataSource {
-    interface Local {
-
-    }
-
-    interface Remote {
-        fun getCovidInformation()
-    }
-}
