--- app/src/main/java/sun/trainingcourse/anticovid_19/data/resource/remote/CovidRemoteDataSource.kt
@@ -1,9 +1,0 @@
-package sun.trainingcourse.anticovid_19.data.resource.remote
-
-import sun.trainingcourse.anticovid_19.data.resource.CovidDataSource
-
-class CovidRemoteDataSource : CovidDataSource.Remote{
-
-    override fun getCovidInformation() {
-    }
-}
