--- app/src/main/java/sun/trainingcourse/anticovid_19/utils/Constants.kt
@@ -1,10 +1,0 @@
-package sun.trainingcourse.anticovid_19.utils
-
-object ModelConst {
-    const val NEW_CONFIRMED = "NewConfirmed"
-    const val NEW_DEATHS = "NewDeaths"
-    const val NEW_RECOVERED = "NewRecovered"
-    const val TOTAL_CONFIRMED = "TotalConfirmed"
-    const val TOTAL_DEATHS = "TotalDeaths"
-    const val TOTAL_RECOVERED = "TotalRecovered"
-}
