+++ app/src/main/res/layout/activity_main.xml
@@ -3,6 +3,6 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.main.MainActivity">
+    tools:context="com.sunasterisk.anticovid_19.ui.main.MainActivity">
 
 </androidx.constraintlayout.widget.ConstraintLayout>
